favorites support multiple users

+++ WoWDatabase/DB/Model/Favorites.swift
@@ -22,6 +22,11 @@ extension Favorites {
 		get { TypeEnum(rawValue: type_ ?? "") ?? .none }
 		set { type_ = newValue.rawValue }
 	}
+	
+	var users: [User] {
+		get { Array((users_ as? Set<User>) ?? []) }
+		set { users_ = Set(newValue) as NSSet }
+	}
 }
 	
 extension Favorites {
@@ -32,28 +37,41 @@ extension Favorites {
 		return request
 	}
 	
-	static func saveFav(achievement: AchievementModel, context: NSManagedObjectContext) -> String {
-		let authService = AuthService(context: context)
-		guard let id = achievement.id, let user = authService.getActiveUser() else { return L10n.General.error }
-		let fav = Favorites(context: context)
-		fav.id = id
-		fav.name = achievement.name
-		fav.type = .achievement
-		fav.userId = NSSet(array: [user])
+	static func saveFav(achievement: AchievementModel, user: User, context: NSManagedObjectContext) -> String {
+		guard let id = achievement.id else { return L10n.General.error }
+		
+		let predicate = NSPredicate(format: "id_ == %@ AND type_ == %@", "\(id)", TypeEnum.achievement.rawValue)
+		let storedFav = try? context.fetch(Favorites.getFavoritesRequest(predicate: predicate)).first
+		
+		if let storedFav = storedFav {
+			storedFav.users.append(user)
+		} else {
+			let fav = Favorites(context: context)
+			fav.id = id
+			fav.name = achievement.name
+			fav.type = .achievement
+			fav.users = [user]
+		}
+		
 		return context.saveContext()
 	}
 	
-	static func removeFav(id: Int, type: TypeEnum, context: NSManagedObjectContext) -> String {
+	static func removeFav(id: Int, type: TypeEnum, user: User, context: NSManagedObjectContext) -> String {
 		let predicate = NSPredicate(format: "id_ == %@ AND type_ == %@", "\(id)", type.rawValue)
 		let fav = try? context.fetch(Favorites.getFavoritesRequest(predicate: predicate)).first
 		guard let fav = fav else { return L10n.General.error }
-		context.delete(fav)
+		
+		fav.users = fav.users.filter { $0 != user }
+		if fav.users.isEmpty {
+			context.delete(fav)
+		}
 		return context.saveContext()
 	}
 	
-	static func checkInFav(id: Int, type: TypeEnum, context: NSManagedObjectContext) -> Bool {
+	static func checkInFav(id: Int, type: TypeEnum, user: User, context: NSManagedObjectContext) -> Bool {
 		let predicate = NSPredicate(format: "id_ == %@ AND type_ == %@", "\(id)", type.rawValue)
 		let achievement = try? context.fetch(Favorites.getFavoritesRequest(predicate: predicate)).first
-		return !(achievement == nil)
+		
+		return achievement?.users.contains { $0 == user } ?? false
 	}
 }
